Update rake && rspec
The version of RSpec being used had bugs preventing properly testing
ReCaptcha. Most impacting was a bug around stubbing module methods
(which is how ReCaptcha provides validation to the controller).

Additionally, the pinned version of RSpec-rails was incompatible
with any newer version of rake. This is likely what was preventing
updating of gems in the first place because all it gave you was a
very unhelpful message about ```undefined method last_comment```

In the future, you should not pin gem versions unless absolutely
necessary as it forces you application to become static over time.